+++ self_play.py
@@ -7,6 +7,8 @@ import torch
 
 import models
 
+from mcts import MCTS, Node, GameHistory, MinMaxStats
+
 
 @ray.remote
 class SelfPlay:
@@ -246,323 +248,3 @@ class SelfPlay:
         return action
 
 
-# Game independent
-class MCTS:
-    """
-    Core Monte Carlo Tree Search algorithm.
-    To decide on an action, we run N simulations, always starting at the root of
-    the search tree and traversing the tree according to the UCB formula until we
-    reach a leaf node.
-    """
-
-    def __init__(self, config):
-        self.config = config
-
-    def run(
-        self,
-        model,
-        observation,
-        legal_actions,
-        to_play,
-        add_exploration_noise,
-        override_root_with=None,
-    ):
-        """
-        At the root of the search tree we use the representation function to obtain a
-        hidden state given the current observation.
-        We then run a Monte Carlo Tree Search using only action sequences and the model
-        learned by the network.
-        """
-        if override_root_with:
-            root = override_root_with
-            root_predicted_value = None
-        else:
-            root = Node(0)
-            observation = (
-                torch.tensor(observation)
-                .float()
-                .unsqueeze(0)
-                .to(next(model.parameters()).device)
-            )
-            (
-                root_predicted_value,
-                reward,
-                policy_logits,
-                hidden_state,
-            ) = model.initial_inference(observation)
-            root_predicted_value = models.support_to_scalar(
-                root_predicted_value, self.config.support_size
-            ).item()
-            reward = models.support_to_scalar(reward, self.config.support_size).item()
-            assert (
-                legal_actions
-            ), f"Legal actions should not be an empty array. Got {legal_actions}."
-            assert set(legal_actions).issubset(
-                set(self.config.action_space)
-            ), "Legal actions should be a subset of the action space."
-            root.expand(
-                legal_actions,
-                to_play,
-                reward,
-                policy_logits,
-                hidden_state,
-            )
-
-        if add_exploration_noise:
-            root.add_exploration_noise(
-                dirichlet_alpha=self.config.root_dirichlet_alpha,
-                exploration_fraction=self.config.root_exploration_fraction,
-            )
-
-        min_max_stats = MinMaxStats()
-
-        max_tree_depth = 0
-        for _ in range(self.config.num_simulations):
-            virtual_to_play = to_play
-            node = root
-            search_path = [node]
-            current_tree_depth = 0
-
-            while node.expanded():
-                current_tree_depth += 1
-                action, node = self.select_child(node, min_max_stats)
-                search_path.append(node)
-
-                # Players play turn by turn
-                if virtual_to_play + 1 < len(self.config.players):
-                    virtual_to_play = self.config.players[virtual_to_play + 1]
-                else:
-                    virtual_to_play = self.config.players[0]
-
-            # Inside the search tree we use the dynamics function to obtain the next hidden
-            # state given an action and the previous hidden state
-            parent = search_path[-2]
-            value, reward, policy_logits, hidden_state = model.recurrent_inference(
-                parent.hidden_state,
-                torch.tensor([[action]]).to(parent.hidden_state.device),
-            )
-            value = models.support_to_scalar(value, self.config.support_size).item()
-            reward = models.support_to_scalar(reward, self.config.support_size).item()
-            node.expand(
-                self.config.action_space,
-                virtual_to_play,
-                reward,
-                policy_logits,
-                hidden_state,
-            )
-
-            self.backpropagate(search_path, value, virtual_to_play, min_max_stats)
-
-            max_tree_depth = max(max_tree_depth, current_tree_depth)
-
-        extra_info = {
-            "max_tree_depth": max_tree_depth,
-            "root_predicted_value": root_predicted_value,
-        }
-        return root, extra_info
-
-    def select_child(self, node, min_max_stats):
-        """
-        Select the child with the highest UCB score.
-        """
-        max_ucb = max(
-            self.ucb_score(node, child, min_max_stats)
-            for action, child in node.children.items()
-        )
-        action = numpy.random.choice(
-            [
-                action
-                for action, child in node.children.items()
-                if self.ucb_score(node, child, min_max_stats) == max_ucb
-            ]
-        )
-        return action, node.children[action]
-
-    def ucb_score(self, parent, child, min_max_stats):
-        """
-        The score for a node is based on its value, plus an exploration bonus based on the prior.
-        """
-        pb_c = (
-            math.log(
-                (parent.visit_count + self.config.pb_c_base + 1) / self.config.pb_c_base
-            )
-            + self.config.pb_c_init
-        )
-        pb_c *= math.sqrt(parent.visit_count) / (child.visit_count + 1)
-
-        prior_score = pb_c * child.prior
-
-        if child.visit_count > 0:
-            # Mean value Q
-            value_score = min_max_stats.normalize(
-                child.reward
-                + self.config.discount
-                * (child.value() if len(self.config.players) == 1 else -child.value())
-            )
-        else:
-            value_score = 0
-
-        return prior_score + value_score
-
-    def backpropagate(self, search_path, value, to_play, min_max_stats):
-        """
-        At the end of a simulation, we propagate the evaluation all the way up the tree
-        to the root.
-        """
-        if len(self.config.players) == 1:
-            for node in reversed(search_path):
-                node.value_sum += value
-                node.visit_count += 1
-                min_max_stats.update(node.reward + self.config.discount * node.value())
-
-                value = node.reward + self.config.discount * value
-
-        elif len(self.config.players) == 2:
-            for node in reversed(search_path):
-                node.value_sum += value if node.to_play == to_play else -value
-                node.visit_count += 1
-                min_max_stats.update(node.reward + self.config.discount * -node.value())
-
-                value = (
-                    -node.reward if node.to_play == to_play else node.reward
-                ) + self.config.discount * value
-
-        else:
-            raise NotImplementedError("More than two player mode not implemented.")
-
-
-class Node:
-    def __init__(self, prior):
-        self.visit_count = 0
-        self.to_play = -1
-        self.prior = prior
-        self.value_sum = 0
-        self.children = {}
-        self.hidden_state = None
-        self.reward = 0
-
-    def expanded(self):
-        return len(self.children) > 0
-
-    def value(self):
-        if self.visit_count == 0:
-            return 0
-        return self.value_sum / self.visit_count
-
-    def expand(self, actions, to_play, reward, policy_logits, hidden_state):
-        """
-        We expand a node using the value, reward and policy prediction obtained from the
-        neural network.
-        """
-        self.to_play = to_play
-        self.reward = reward
-        self.hidden_state = hidden_state
-
-        policy_values = torch.softmax(
-            torch.tensor([policy_logits[0][a] for a in actions]), dim=0
-        ).tolist()
-        policy = {a: policy_values[i] for i, a in enumerate(actions)}
-        for action, p in policy.items():
-            self.children[action] = Node(p)
-
-    def add_exploration_noise(self, dirichlet_alpha, exploration_fraction):
-        """
-        At the start of each search, we add dirichlet noise to the prior of the root to
-        encourage the search to explore new actions.
-        """
-        actions = list(self.children.keys())
-        noise = numpy.random.dirichlet([dirichlet_alpha] * len(actions))
-        frac = exploration_fraction
-        for a, n in zip(actions, noise):
-            self.children[a].prior = self.children[a].prior * (1 - frac) + n * frac
-
-
-class GameHistory:
-    """
-    Store only usefull information of a self-play game.
-    """
-
-    def __init__(self):
-        self.observation_history = []
-        self.action_history = []
-        self.reward_history = []
-        self.to_play_history = []
-        self.child_visits = []
-        self.root_values = []
-        self.reanalysed_predicted_root_values = None
-        # For PER
-        self.priorities = None
-        self.game_priority = None
-
-    def store_search_statistics(self, root, action_space):
-        # Turn visit count from root into a policy
-        if root is not None:
-            sum_visits = sum(child.visit_count for child in root.children.values())
-            self.child_visits.append(
-                [
-                    root.children[a].visit_count / sum_visits
-                    if a in root.children
-                    else 0
-                    for a in action_space
-                ]
-            )
-
-            self.root_values.append(root.value())
-        else:
-            self.root_values.append(None)
-
-    def get_stacked_observations(self, index, num_stacked_observations):
-        """
-        Generate a new observation with the observation at the index position
-        and num_stacked_observations past observations and actions stacked.
-        """
-        # Convert to positive index
-        index = index % len(self.observation_history)
-
-        stacked_observations = self.observation_history[index].copy()
-        for past_observation_index in reversed(
-            range(index - num_stacked_observations, index)
-        ):
-            if 0 <= past_observation_index:
-                previous_observation = numpy.concatenate(
-                    (
-                        self.observation_history[past_observation_index],
-                        [
-                            numpy.ones_like(stacked_observations[0])
-                            * self.action_history[past_observation_index + 1]
-                        ],
-                    )
-                )
-            else:
-                previous_observation = numpy.concatenate(
-                    (
-                        numpy.zeros_like(self.observation_history[index]),
-                        [numpy.zeros_like(stacked_observations[0])],
-                    )
-                )
-
-            stacked_observations = numpy.concatenate(
-                (stacked_observations, previous_observation)
-            )
-
-        return stacked_observations
-
-
-class MinMaxStats:
-    """
-    A class that holds the min-max values of the tree.
-    """
-
-    def __init__(self):
-        self.maximum = -float("inf")
-        self.minimum = float("inf")
-
-    def update(self, value):
-        self.maximum = max(self.maximum, value)
-        self.minimum = min(self.minimum, value)
-
-    def normalize(self, value):
-        if self.maximum > self.minimum:
-            # We normalize only when we have set the maximum and minimum values
-            return (value - self.minimum) / (self.maximum - self.minimum)
-        return value
